orchestrators/bidding-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
httpx==0.27.0
aio-pika==9.4.3
temporalio==1.7.1
grpcio==1.66.0
grpcio-tools==1.66.0
python-dotenv==1.0.1
strawberry-graphql[fastapi]==0.243.1
